# docs/api/openapi.yaml
openapi: 3.1.0 # または 3.0.3 など広くサポートされているバージョン
info:
  title: TASK MVP API
  version: 0.1.0 # プロジェクトのバージョンに合わせて更新
  description: AI秘書「TASK」のバックエンドAPI仕様書 (MVP)
servers:
  - url: /api/v1 # APIのベースパス
    description: アプリケーションルートからの相対パス
components:
...
  responses:
    NoContent:
      description: 成功（返却データなし）
      content: {}
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: エラーメッセージ
            required: [message]
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: エラーメッセージ
            required: [message]
    Forbidden:
      description: アクセス権限なし
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

  # --- 再利用可能なデータスキーマ定義 ---
  schemas:
...
    CreateScheduleInput:
      type: object
      properties:
        title:
          type: string
          description: 日程調整のタイトル
          minLength: 1
        description:
          type: string
          description: 詳細説明 (任意)
          nullable: true
      required:
        - title
...
    FirestoreTimestamp: # Timestampの表現 (共通化)
      type: string
      format: date-time
      description: ISO8601形式の日時文字列 (UTC)
      example: "2025-05-02T10:50:56.123Z"
...
    Schedule:
      type: object
      properties:
        id: { type: string, description: Schedule ID, readOnly: true }
        title: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [adjusting, confirmed, cancelled] }
        createdBy: { type: string, description: User ID, readOnly: true }
        groupId: { type: string, nullable: true }
        createdAt: { $ref: '#/components/schemas/FirestoreTimestamp' }
        updatedAt: { $ref: '#/components/schemas/FirestoreTimestamp' }
        candidateDates: # 候補日リスト
          type: array
          items:
            type: object
            properties:
              optionId: { type: string }
              datetime: { $ref: '#/components/schemas/FirestoreTimestamp' }
            required: [optionId, datetime]
          nullable: true
        confirmedDateTime: { $ref: '#/components/schemas/FirestoreTimestamp', nullable: true }
        isArchived: { type: boolean, default: false, readOnly: true }
      required: [id, title, status, createdBy, createdAt, updatedAt, isArchived]
...
    SharedMemo:
      type: object
      description: 共有メモのデータ構造
      properties:
        id: { type: string, description: Shared Memo ID, readOnly: true }
        title: { type: string, minLength: 1 }
        templateType: { type: string, enum: [meeting, outing, free], default: free }
        content: { type: string, nullable: true }
        createdBy: { type: string, description: Creator User ID, readOnly: true }
        groupId: { type: string, nullable: true }
        createdAt: { $ref: '#/components/schemas/FirestoreTimestamp' }
        updatedAt: { $ref: '#/components/schemas/FirestoreTimestamp' }
        lastEditorId: { type: string, nullable: true, readOnly: true }
        readableUserIds:
          type: array
          items: { type: string }
          description: List of user IDs who can read/write this memo
          readOnly: true
        isArchived: { type: boolean, default: false, readOnly: true }
      required: [id, title, createdBy, createdAt, updatedAt, readableUserIds, isArchived]
...
    CreateSharedMemoInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        templateType: { type: string, enum: [meeting, outing, free], nullable: true }
        content: { type: string, nullable: true }
        groupId: { type: string, nullable: true, description: 関連付けるグループID (任意) }
      required: [title]
    UpdateSharedMemoInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        content: { type: string, nullable: true }
...
    ListSharedMemosResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/SharedMemo' }
        nextCursor: { type: string, nullable: true }
      required: [data]

    # --- Common Schemas ---
    ErrorResponse:
      type: object
      properties:
        error: { type: string, description: エラーメッセージ }
      required: [error]

  # --- 認証方式の定義 ---
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }

# --- デフォルトで全APIに認証を適用 ---
security:
  - bearerAuth: []

# --- APIパス (エンドポイント) 定義 ---
paths:
...
  /schedules:
    post:
      summary: 新規スケジュール作成
      operationId: createSchedule
      tags: [Schedule]
      security: [{"bearerAuth": []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateScheduleInput' }
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /shared-memos:
    post:
      summary: Create a shared memo
      operationId: createSharedMemo
      tags: [Shared Memo]
      security: [{"bearerAuth": []}]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateSharedMemoInput' }}}
      responses:
        '201':
          description: Shared memo created
          content: { application/json: { schema: { $ref: '#/components/schemas/SharedMemo' }}}
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      summary: List accessible shared memos (pagination)
      operationId: listSharedMemos
      tags: [Shared Memo]
      security: [{"bearerAuth": []}]
      parameters:
        - name: limit
          in: query
          description: Max number of memos to return
          required: false
          schema: { type: integer, maximum: 50, default: 20 }
        - name: cursor
          in: query
          description: Cursor for pagination (last memo ID from previous page)
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Success
          content: { application/json: { schema: { $ref: '#/components/schemas/ListSharedMemosResponse' }}}
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /shared-memos/{id}:
    get:
      summary: Get a shared memo by ID
      operationId: getSharedMemoById
      tags: [Shared Memo]
      security: [{"bearerAuth": []}]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: Shared Memo ID }
      responses:
        '200':
          description: Success
          content: { application/json: { schema: { $ref: '#/components/schemas/SharedMemo' }}}
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update a shared memo
      operationId: updateSharedMemo
      tags: [Shared Memo]
      security: [{"bearerAuth": []}]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: Shared Memo ID }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateSharedMemoInput' }}}
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Archive (soft delete) a shared memo
      operationId: archiveSharedMemo
      tags: [Shared Memo]
      security: [{"bearerAuth": []}]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: Shared Memo ID }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...